Corregidos algunos fallos en la tarea de actualización de usuarios en modo examen.

+++ classes/task/update_exammode_users.php
@@ -24,6 +24,9 @@
  */
 namespace local_exammode\task;
 
+use local_exammode\objects\exammode;
+use local_exammode\objects\exammode_user;
+
 class update_exammode_users extends \core\task\scheduled_task {
 
     public function get_name() {
@@ -33,10 +36,16 @@ class update_exammode_users extends \core\task\scheduled_task {
     public function execute() {
         $manager = \local_exammode\manager::get_instance();
 
+        // Courses that should be in exammode must not remove users aunque
+        // it has exammodes to remove.
+        $courses_to_skip = array();
+
         // Courses that should be in exammode.
         $courses = $manager->get_courses_in_exammode();
         foreach ($courses as $exammode) {
             try {
+                $courses_to_skip[] = $exammode->get_courseid();
+
                 if ($exammode->get_state() === \local_exammode\objects\exammode::STATE_PENDING) {
                     $exammode->set_state(\local_exammode\objects\exammode::STATE_CONFIGURING);
                     $manager->update_exam($exammode);
@@ -59,17 +68,29 @@ class update_exammode_users extends \core\task\scheduled_task {
         $exammodes = $manager->get_finished_exammodes();
         foreach ($exammodes as $exammode) {
             try {
-                if ($exammode->get_state() !== \local_exammode\objects\exammode::STATE_UNCONFIGURING) {
-                    $exammode->set_state(\local_exammode\objects\exammode::STATE_UNCONFIGURING);
+                // Skip courses that are in exammode now.
+                if (in_array($exammode->get_courseid(), $courses_to_skip)) {
+                    continue;
+                }
+
+                if ($exammode->get_state() !== exammode::STATE_UNCONFIGURING &&
+                        $exammode->get_state() !== exammode::STATE_TODELETE ) {
+
+                    $exammode->set_state(exammode::STATE_UNCONFIGURING);
                     $manager->update_exam($exammode);
                 }
 
                 $users = $manager->get_users_in_exammode($exammode->get_courseid());
                 foreach ($users as $emu) {
                     $manager->remove_user_from_exammode($emu);
+
+                    \local_exammode\event\user_abandoned::create_from_exammode($exammode, $emu)->trigger();
                 }
 
-                $exammode->set_state(\local_exammode\objects\exammode::STATE_FINISHED);
+                if ($exammode->get_state() !== exammode::STATE_TODELETE) {
+                    $exammode->set_state(\local_exammode\objects\exammode::STATE_FINISHED);
+                }
+
                 $manager->update_exam($exammode);
             } catch (\Exception $e) {
                 mtrace("        ERROR: " . $e->getMessage() . ": " . $e->getTraceAsString());
@@ -80,12 +101,20 @@ class update_exammode_users extends \core\task\scheduled_task {
         $exammodes = $manager->get_exammodes_to_delete();
         foreach ($exammodes as $em) {
             try {
+                // Skip courses that are in exammode now.
+                if (in_array($em->get_courseid(), $courses_to_skip)) {
+                    continue;
+                }
+
                 $users = $manager->get_users_in_exammode($em->get_courseid());
                 foreach ($users as $emu) {
                     $manager->remove_user_from_exammode($emu);
+                    \local_exammode\event\user_abandoned::create_from_exammode($em, $emu)->trigger();
                 }
 
                 $manager->delete_exam($em->get_id());
+
+                \local_exammode\event\exam_deleted::create_from_exammode($em)->trigger();
             } catch (\Exception $e) {
                 mtrace("        ERROR: " . $e->getMessage() . ": " . $e->getTraceAsString());
             }
@@ -107,15 +136,15 @@ class update_exammode_users extends \core\task\scheduled_task {
         foreach ($current as $cu) {
             $found = false;
             foreach ($potential as $au) {
-                if ($cu->get_exammodeid() == $au->get_exammodeid()) {
-                    if ($cu->get_userid() == $au->get_userid()) {
-                        $found = true;
-                        break;
-                    }
+                if ($cu->get_userid() == $au->get_userid()) {
+                    $found = true;
+                    break;
                 }
             }
             if (!$found) {
+                mtrace("        Remove user " . $cu->get_userid() . " from exammode " . $cu->get_exammodeid(), false);
                 $manager->remove_user_from_exammode($cu);
+                mtrace("OK");
 
                 $event = \local_exammode\event\user_abandoned::create_from_exammode($exammode, $cu);
                 $event->trigger();
@@ -140,19 +169,20 @@ class update_exammode_users extends \core\task\scheduled_task {
         foreach ($potential as $au) {
             $found = false;
             foreach ($current as $cu) {
-                if ($cu->get_exammodeid() == $au->get_exammodeid()) {
-                    if ($cu->get_userid() == $au->get_userid()) {
-                        $found = true;
-                        break;
-                    }
+                if ($cu->get_userid() == $au->get_userid()) {
+                    $found = true;
+                    break;
                 }
             }
             if (!$found) {
                // OJO. Es necesario ejecutar my_copy_page para asegurar que existen las
                // instancias de bloque en la tabl block_instances para el dashboard.
                 require_once($CFG->dirroot . '/my/lib.php');
-                if ( \my_copy_page($au->get_userid()) !== false) {
+                if (\my_copy_page($au->get_userid()) !== false) {
+                    mtrace("        Add user " . $au->get_userid() . " to exammode " . $au->get_exammodeid() . "... ", false);
                     $manager->put_user_in_exammode($au);
+                    mtrace("OK");
+
 
                     $event = \local_exammode\event\user_started::create_from_exammode($exammode, $au);
                     $event->trigger();
